Refactor: move conditional field logic to RenderDynamicFields instead of field itself
