+++ src/app/[locale]/incident/add/components/questions/RadioInput.tsx
@@ -1,76 +1,20 @@
-import React, { useEffect, useMemo, useState } from 'react'
+import React from 'react'
 import { useTranslations } from 'next-intl'
-import { useFormStore } from '@/store/form_store'
 import { QuestionField } from '@/types/form'
 import { getValidators } from '@/lib/utils/form-validator'
 import { useFormContext } from 'react-hook-form'
-import { evaluateConditions } from '@/lib/utils/check-visibility'
 import { Paragraph } from '@/components/index'
 
 interface RadioGroupProps extends QuestionField {}
 
 export const RadioInput = ({ field }: RadioGroupProps) => {
-  const [shouldRender, setShouldRender] = useState<boolean>(false)
   const {
-    watch,
     register,
     formState: { errors },
-    setValue,
   } = useFormContext()
   const t = useTranslations('general.errors')
-  const { formState: formStoreState } = useFormStore()
 
   const errorMessage = errors[field.key]?.message as string
-
-  const watchValues = watch()
-
-  // Memoize `evaluateConditions` result to prevent unnecessary updates
-  const shouldRenderResult = useMemo(
-    () => evaluateConditions(field.meta, watchValues),
-    [field.meta, watchValues]
-  )
-
-  // Handle visibility changes
-  useEffect(() => {
-    if (shouldRender !== shouldRenderResult) {
-      setShouldRender(shouldRenderResult)
-      if (!shouldRenderResult) {
-        setValue(field.key, null)
-      } else {
-        const defaultValue = getDefaultValueRadioInput(field.key)
-        if (defaultValue) {
-          setValue(field.key, defaultValue)
-        }
-      }
-    }
-  }, [shouldRenderResult, shouldRender, field.key, setValue])
-
-  // Get default value helper function
-  const getDefaultValueRadioInput = (id: string) => {
-    const extraProperty = formStoreState.extra_properties.find(
-      (question) => question.id === id
-    )
-
-    if (!extraProperty) return null
-
-    if (typeof extraProperty.answer !== 'string' && extraProperty.answer?.id) {
-      return extraProperty.answer.id
-    }
-
-    return null
-  }
-
-  // Register the field immediately with initial value
-  useEffect(() => {
-    const defaultValue = getDefaultValueRadioInput(field.key)
-    if (defaultValue && shouldRender) {
-      setValue(field.key, defaultValue)
-    }
-  }, [field.key, setValue, shouldRender])
-
-  if (!shouldRender) {
-    return null
-  }
 
   return (
     <fieldset aria-invalid={!!errorMessage} role="radiogroup">
