+++ src/services/additional-questions/index.ts
@@ -1,10 +1,14 @@
 import { signalsClient } from '@/services/client/api-client'
+import { PublicQuestion } from '@/types/form'
 
-export const fetchAdditionalQuestions = async (main: string, sub: string) => {
+export const fetchAdditionalQuestions = async (
+  main: string,
+  sub: string
+): Promise<PublicQuestion[]> => {
   try {
     const data = await signalsClient.v1.v1PublicQuestionsList(main, 1, 100, sub)
 
-    const questions = data.results
+    const questions = data.results as unknown as PublicQuestion[]
     if (!questions) {
       return []
     }
